Add Travis.yml stub
Commit adds Travis.yml file that will build the static files for the
website and upload them to the server. It installs hugo, go, s3cmd and a
few more dependencies. After building and verifying the build, the
static html and css files are uploaded via s3cmd.

Currently the script and deploy stages are commented out as the name of
the website and the domain for the server are not yet decided upon.