Add Dockerfile and new theme, update CD
Added Dockerfile for project to run Hugo inside Docker container instead
of barebone. This way, we can run a Docker container inside Kubernetes
instead of provisioning a separate server for the website only.

After successful Hugo build, the image will be pushed to
kk-digital/website and from there can be deploted/auto-redeployed to
Kubernetes cluster or server of our choice.